[2296] Create a dedicated hook for the export to image action
Also move the code for diagram panel in a dedicated folder

Bug: https://github.com/eclipse-sirius/sirius-web/issues/2296
