+++ packages/diagrams/frontend/sirius-components-diagrams-reactflow/src/renderer/panel/DiagramPanel.types.ts
