+++ packages/sirius-web/frontend/sirius-web/package.json
@@ -33,6 +33,7 @@
     "d3": "7.0.0",
     "fontsource-roboto": "4.0.0",
     "graphql": "16.8.0",
+    "html-to-image": "1.11.11",
     "notistack": "3.0.1",
     "prop-types": "15.8.1",
     "react": "17.0.2",
